+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Make group event to have the same UI as regular event [#3914](https://github.com/rokwire/illinois-app/issues/3914).
 ### Added
 - Show published status in event detail panel for admins only [#3908](https://github.com/rokwire/illinois-app/issues/3908).
+- Added published info link in Events filters panel [#3886](https://github.com/rokwire/illinois-app/issues/3886).
 
 ## [5.1.17] - 2024-01-30
 ### Added